requirements.txt
streamlit
google-generativeai
python-dotenv
openai-agents  # If using OpenAI Agent SDK
